requirements.txt
mss==7.0.1
opencv_python==4.6.0.66
Pillow==10.2.0
pyserial==3.5
streamlit==1.31.0
vision==1.0.0
